fix: split Route into own file, fix ktlint blank lines
- Extract sealed class Route into Route.kt (fixes detekt MatchingDeclarationName)
- Add blank lines between Route members (fixes ktlint spacing)
- Keep Navigation.kt with PortalUsuarioNavHost only
